Move Electron duplicate modifier check to `BaseResolvedKeybinding`
